Add support for scripts created with eval
This commit add support for scripts created by eval, by adding a
secondary cache.

The cache maps from script_ids to URL. It is build by parsing the
Debugger.* messages, which contain a stacktrace for most scripts.

The cache is then used to lookup URLs for scripts which do not have a
URL.

This commit also updates the message parsing. It combines the functions
find_node and add_dependency, because they were always used as a pair
and this simplifies handling of scripts by script_id.