Refactor new() and process_messages()
New now requires the list of messages. This is because calling
simplify_graph() and then process_messages() again would be invalid.
This is because removing nodes changes the node IDs, thus the node ID
cache is invalid after simplify_graph(). Prevent this problem by only
allowing to call process_messages() once. This removes the need for the
node_cache in the DepGraph struct.

Do not process data URIs any longer. This works by making create_node
return a Result<Option<NodeIndex>, _> instead of Result<NodeIndex, _>
such that for data URIs None is returned.