Fix dailymotion.com crashes
It seems like dailymotion changes the page website and append /de to it,
without it being a network request (so with JS).
This causes the case, that the website URL is unknown and thus in the
cases where a node for it should exists, it crashes.

Fix it by listening for Target.targetInfoChanged events, which at least
tell whenever the currect website URL changes.

Fixes #8