Add initial support for confusion domains
You can specify on the command line multiple files which each are CSV
files mapping from a domain to a more canonical domain. The files are
processed in order. Duplicate entries are only allowed it they are fully
identical. Two mappings for the same domain with different canonical
domains is forbidden.

This provides a mechanism to support redirects and other "identical"
domains.

The problem with it is that it favours those domains where many map to a
single canonical representation, as they are overrepresented in the
dataset.