Increase performance of Edit Distance with an early abort
Most distance comparisons of Sequences are uninteresting, as the values
will be larger than the current best values, thus they can never replace
the current values. By specifying a maximal distance which we care about,
we can abort early, if we see that the distance will definitely be larger
than the maximal distance thus avoiding extra work. An early abort will
return a distance which is larger than the maximal distance, in order to
ensure that the entry will never be considered as "better".

In the `take_smallest` function we now pass the currently largest
distance we track to the `distance` function, thus enabling the early
abort.

Measuring the CPU time with `time` shows clear performance benefits:

OLD:
        Command being timed: "./dns-sequence-compare-old --misclassifications=/dev/null -k 1 ../../dnscaptures_debug/"
        User time (seconds): 129.76
        System time (seconds): 0.22
        Percent of CPU this job got: 267%
        Elapsed (wall clock) time (h:mm:ss or m:ss): 0:48.65

	Command being timed: "./dns-sequence-compare-old --misclassifications=/dev/null -k 1 ../../dnscaptures_debug/"
	User time (seconds): 129.59
	System time (seconds): 0.22
	Percent of CPU this job got: 268%
	Elapsed (wall clock) time (h:mm:ss or m:ss): 0:48.34

NEW:
        Command being timed: "./dns-sequence-max-distance --misclassifications=/dev/null -k 1 ../../dnscaptures_debug/"
        User time (seconds): 78.88
        System time (seconds): 0.18
        Percent of CPU this job got: 266%
        Elapsed (wall clock) time (h:mm:ss or m:ss): 0:29.70

	Command being timed: "./dns-sequence-max-distance --misclassifications=/dev/null -k 1 ../../dnscaptures_debug/"
	User time (seconds): 78.92
	System time (seconds): 0.21
	Percent of CPU this job got: 267%
	Elapsed (wall clock) time (h:mm:ss or m:ss): 0:29.60