Implement better classification statistics
Instead of only providing the three steps "Correct", "Undecided",
"Wrong" provide much more detailed information.
The new hierarchy is "Exact", "Majority", "Plurality",
"PluralityThenMinDist", "Contains", "Wrong". Each level (with the
exception of "Wrong") refines the previous level and is a direct subset.
For example: Every Majority is also a Plurality and every Exact is also
a Majority.

The misclassifications output got changed to JSON, because CSV is not
able to represent the nested data structures.