Be more clever about handling xvfb
Keeping the same xvfb instance for all times does not work, as it seems
to sometimes fail and then the docker/chrome also gets stuck.
Circumventing this by spawning a new xvfb for each tasks works, but is
wastefull.

Instead keep a single xvfb for multiple tasks but restart it if:

* The task fails for any reason, including timeouts which could be
caused by a not running xvfb.
* If the process status of xvfb indicates it is not running anymore or
there are errors in retriving the status. Basically anything which is
not definitely "yes, it is running".