Add new taskmanager subcommand
Add subcommand `add`

It is very similar to `init`, but it does not empty the database.
It is intended to be repeatedly called on an existing database, to add
new, higher priority tasks.

As such, it is ignoring the `initial_priority` value and uses
`per_domain_datasets_repeated_measurements` instead of
`per_domain_datasets`.