Fix Not all docker containers are stopped properly
Symptom:
After gathering traces some docker contains keep running and
are not stopped. This is likely also the cause for the +100 load average
during the weekend.

What happened:
The `timeout` does not actually stop the docker container. While it is
not clear to me why this is the case, it only stops the interation with
the contanier, but the containers keep running. As such, they keep
accumulating and could bring down the host system.

Docker does not provide a way to be combined with a wall clock timeout.

Solution:
We wait on the docker process with a timeout in the Rust part. If the
docker process did not finish until the timeout is reached, we try to
make really sure, that we stop the docker container. We use `docker
kill` and `docker rm --force=true` in order to force the docker
container to stop and delete the container.

There is a Rust crate to wait on a process with a timeout. It was not
updated since last year and there are outstanding bug fixes in the
repository, which are not yet released on crates.io. As such, we use the
github version of the library.