Fix wrong classifications in `classify` setting
The test data was not subject to the confusion domain mappings,
therefore even a correct match could have the wrong label, as the
trainings data label was renamed, but the test data label was not.

The problem was that the confusion domains mapping was cleared from
memory, when it was still needed.
This is not a problem for crossvalidation, as it takes the test data
from the trainings data and the trainings data is correctly labelled.