Update pcap-parsing to support multiple streams per pcap
The timeouts seen in reality are too short such that a single website
will likely be loaded over multiple TCP streams.

For this we need to track multiple TCP streams by both source and
destination parameters. We can filter each stream separatly, with the
code we already have.

Update tests.