Move the core logig of extract-sequence to sequence
Sequence::from_path should be able to read dnstap, json, and pcap files.
Since the logic for reading pcap files was encapsulated in
extract-sequence, we need to make this accessible from sequence.
However, this would lead to a circular dependency.

Therefore, it is necessary, to move the code to the sequence crate.
The new code is optional behind a feature flag, as it depends on so many
other dependencies.

The extract-sequence now depends on sequence with the feature.