Switch from Chrome to Chromium Headless
Chromium offers the same devtools protocol as Chrome and is therefore
a simple replacement.
It may be a bit better with not sending data to different Google domains
which sometimes introduced wrong requests. But this is not tested so
far. It also simplifies the docker file a bit since Chromium can be
directly installed and no X relics are necessary.