Fix reading of PCAPs
After changing the pcap parser, there was an offset too much while
processing packets.
The crate comes with a function which properly removes the pcap header
and returns the data for the inner packet. However, in the case of Linux
cooked capture (L3) before it was necessary to do by hand, but now this
is done by the crate.

Add test that this does not happen in the future.