Add test for pcap with reordered segments
For this we also need to fix the usage of FlowId.
Currently the use was mixed between identifying a two-way flow (sorted
Tuples) and one-way flows (not-sorted, source/destination order
matters). Split them into two such that they cannot be confused.

The problem is, if a two-way flow only uses one buffer. In this case the
bytes of client and server would be mixed and could lead to parsing
errors.