Improve kNN performance by memorizing distances
We introduce a anonymized version of sequences (no ID) such that we can
store them in a map and deduplicate results. For this we need to intern
the Vec<SequenceElement> entries such that they do not blow up the
memory requirements when stored in the map and that we have no lifetime
issues, since we cannot borrow from the sequences.
We also use pre-calculated hash values for the interned sequences to
avoid walking over the vecs every time.

In the kNN part we extract the distance calculation into its own
function. Since we pre-calculate the results now, we need to remove the
early stop optimization of stopping early when the distance from the
Levenshtein distance is definitly larger than the current best
neighbors.
This is necessary as we do not know what the best neighbors are in all
cases.
This code does not fully remove this code part yet, but does no longer
rely on it.