Fix parsing of pcaps with TLS 1.2
Cloudlfare only seems to apply padding for connections using TLS 1.3.  A
small fraction of connections (1160) was seen making connections using
TLS 1.2 though.  This breaks our extraction mechanism which normally
looks ony at messages with at least 128 bytes.

In order to solve the problem, we check the ServerHello handshake
message, which TLS version the server picked and only apply the
filtering if we know it is TLS 1.3.

Curiously, the small fraction of connections only affected these domains
* facebook.com
* goo.gl
* tumblr.com
* wikipedia.com
* wikipedia.org
* wordpress.com