Split dnstap stuff from load_sequences into own module
* The handling of dnstap files is now encapsulated into the dnstap
module. However, it is public, such that other parts can use it. For
example, the plot-dnstap binary.
* Improve documentation for dnstap module and name functions analogue to
the pcap module.
* Remaining parts of load_sequences are the convert_to_* functions and
the handling of gap and padding sizes.

* Remove all dnstiming chart plotting code from encrypted-dns binary and
corresponding python script. Most of the functionality is superseeded
by the plot-dnstap binary. Additionally, the handling of dnstap files
was mostly similar to the sequences::dnstap module, yet slightly
incompalible, such that porting it is infeasible.